debug

+++ src/msp/bitaflught_msp.cpp
@@ -67,8 +67,10 @@ bool BitaflughtMsp::recv(std::uint8_t *command_id, void *payload,
       if (checksumCalc == checksum) {
         return true;
       } else {
-        std::cout << "Invalid checksum(clalc): " << checksumCalc << std::endl;
-        std::cout << "Invalid checksum(recv): " << checksum << std::endl;
+        std::cout << "Invalid checksum(clalc): "
+                  << static_cast<int>(checksumCalc) << std::endl;
+        std::cout << "Invalid checksum(recv): " << static_cast<int>(checksum)
+                  << std::endl;
       }
     }
   }
